+++ packages/xchain-bitcoin/src/index.ts
@@ -1,4 +1,6 @@
 export * from './types'
-export * from './client'
+export { ClientKeystore as Client } from './clientKeystore'
+export { ClientLedger } from './clientLedger'
+export { defaultBTCParams } from './client'
 export * from './const'
 export { getPrefix, validateAddress } from './utils'