+++ packages/xchain-mayachain-amm/src/types.ts
@@ -1,0 +1,4 @@
+export type TxSubmitted = {
+  hash: string
+  url: string
+}