+++ packages/xchain-client/CHANGELOG.md
@@ -1,8 +1,14 @@
-# v0.15.5 (2023-12-01)
+# v0.15.6 (2023-12-11)
 
 ## Update
 
 - UTXO client removed
+
+# v0.15.5 (2023-12-06)
+
+## Add
+
+- Expose Ledger implementation for Bitcoin
 
 # v0.15.4 (2023-11-21)