+++ packages/xchain-mayachain-query/CHANGELOG.md
@@ -1,8 +1,21 @@
-# v0.1.1 (2023-12-01)
+# v0.1.3 (2023-12-11)
 
 ## Update
 
 - Client dependency updated
+
+# v0.1.2 (2023-12-11)
+
+## Update
+
+- Client dependency increased
+
+# v0.1.1 (2023-12-06)
+
+## Update
+
+- Decimals bug fix with Cacao in quote swap
+- Bug fix in getChainInboundDetails fixed
 
 # v0.1.0-alpha (2023-08-03)