+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,8 +1,14 @@
-# v0.1.1 (2023-12-01)
+# v0.1.2 (2023-12-11)
 
 ## Update
 
 - Client dependency updated
+
+# v0.1.1 (2023-12-11)
+
+## Update
+
+- Client dependency increased
 
 # v0.1.0 (2023-09-24)